refactor: name change from CodeV2 to CodeEditv2
