+++ packages/blend/lib/components/CodeEditorV2/CodeEditorV2Header.tsx
@@ -2,7 +2,7 @@ import { Check, Copy, FileCode } from 'lucide-react'
 import Block from '../Primitives/Block/Block'
 import Button from '../Button/Button'
 import { ButtonSize, ButtonSubType, ButtonType } from '../Button/types'
-import { CodeV2HeaderProps } from './codeV2.types'
+import { CodeV2HeaderProps } from './codeEditorV2.types'
 
 export const CodeV2Header = ({
     header,
