+++ packages/blend/lib/components/CodeEditorV2/MonacoEditor/monaco-editor.css
