+++ packages/blend/lib/components/CodeEditorV2/MonacoEditor/monacoTheme.ts
@@ -1,5 +1,5 @@
 import type * as Monaco from 'monaco-editor'
-import type { CodeV2Tokens } from '../codeV2.tokens'
+import type { CodeEditorV2Tokens } from '../codeEditorV2.tokens'
 const MONACO_COLOR_FALLBACK = '#24292F'
 
 const toMonacoColor = (
@@ -10,7 +10,7 @@ const toMonacoColor = (
 /** Token names mapped to syntax key and optional font style for Monaco theme rules */
 const SYNTAX_RULE_SPEC: Array<{
     token: string
-    syntaxKey: keyof CodeV2Tokens['body']['syntax']
+    syntaxKey: keyof CodeEditorV2Tokens['body']['syntax']
     fontStyle?: string
 }> = [
     { token: 'keyword', syntaxKey: 'keyword', fontStyle: 'bold' },
@@ -54,7 +54,7 @@ const SYNTAX_RULE_SPEC: Array<{
 ]
 
 export function createEditorTheme(
-    tokens: CodeV2Tokens
+    tokens: CodeEditorV2Tokens
 ): Monaco.editor.IStandaloneThemeData {
     const syntax = tokens.body.syntax
     const gutter = tokens.body.gutter
