+++ packages/blend/lib/components/CodeEditorV2/index.ts
@@ -1,2 +1,2 @@
 export { default as CodeEditorV2 } from './CodeEditorV2'
-export * from '../codeV2.types'
+export * from './codeEditorV2.types'
